make:crud * 4

+++ src/Controller/CompanyController.php
@@ -3,54 +3,46 @@
 namespace App\Controller;
 
 use App\Entity\Company;
-use App\Entity\TypeOfCompany;
-use App\Form\CompaniesType;
-use App\Form\TypeOfCompaniesType;
+use App\Form\CompanyType;
 use App\Repository\CompanyRepository;
-use App\Repository\TypeOfCompanyRepository;
 use Doctrine\ORM\EntityManagerInterface;
-use Symfony\Component\HttpFoundation\Request;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
 
+#[Route('/company')]
 final class CompanyController extends AbstractController
 {
-    #[Route('/company', name: 'app_company')]
-    public function company(CompanyRepository $companyRepo): Response
+    #[Route(name: 'app_company_index', methods: ['GET'])]
+    public function index(CompanyRepository $companyRepository): Response
     {
-        $companies = $companyRepo->findAll();
-
-        return $this->render('company/index.html.twig', [ 
-            'controller_name' => 'HomePageController',
-            'companies' => $companies,
+        return $this->render('company/index.html.twig', [
+            'companies' => $companyRepository->findAll(),
         ]);
     }
 
-    #[Route('/companyNew', name: 'app_company_new')]
-    public function addCompany(Request $request, EntityManagerInterface $entityManager): Response
+    #[Route('/new', name: 'app_company_new', methods: ['GET', 'POST'])]
+    public function new(Request $request, EntityManagerInterface $entityManager): Response
     {
         $company = new Company();
-
-        $form = $this->createForm(CompaniesType::class, $company);
+        $form = $this->createForm(CompanyType::class, $company);
         $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()){ 
+        if ($form->isSubmitted() && $form->isValid()) {
             $entityManager->persist($company);
-            $entityManager->flush(); 
+            $entityManager->flush();
 
-            $this->addFlash('success', 'The new company have been added !');
-
-            return $this->redirectToRoute('app_company'); 
+            return $this->redirectToRoute('app_company_index', [], Response::HTTP_SEE_OTHER);
         }
 
-        return $this->render('company/addCompany.html.twig', [ // a modifier
-            'controller_name' => 'HomePageController',
-            'formCompany'=> $form->createView(),
+        return $this->render('company/new.html.twig', [
+            'company' => $company,
+            'form' => $form,
         ]);
     }
 
-    #[Route('/company/{id}', name: 'app_company_show', methods: ['GET'])]
+    #[Route('/{id}', name: 'app_company_show', methods: ['GET'])]
     public function show(Company $company): Response
     {
         return $this->render('company/show.html.twig', [
@@ -58,71 +50,32 @@ final class CompanyController extends AbstractController
         ]);
     }
 
-    #[Route('/companyDelete/{id}', name: 'app_company_delete')]
-    public function deleteCompanies(Company $company, EntityManagerInterface $entityManager): Response
+    #[Route('/{id}/edit', name: 'app_company_edit', methods: ['GET', 'POST'])]
+    public function edit(Request $request, Company $company, EntityManagerInterface $entityManager): Response
     {
-        $entityManager->remove($company);
-        $entityManager->flush(); 
-
-        $this->addFlash('danger', 'The company have been deleted.');
-
-        return $this->redirectToRoute('app_company'); 
-        
-    }
-// ---------------------------------------------------------------------------------------------------------------------------------
-#region TYPE OF
-    #[Route('/company/typeOf', name: 'app_company_typeOf')]
-    public function companyTypeOf(TypeOfCompanyRepository $typeOfRepo): Response
-    {
-        $typesOf = $typeOfRepo->findAll();
-
-        return $this->render('company/TypeOfCompanies.html.twig', [ 
-            'controller_name' => 'HomePageController',
-            'types' => $typesOf,
-        ]);
-    }
-
-    #[Route('/company/typeOfNew', name: 'app_company_typeOf_new')]
-    public function addCompanyTypeOf(Request $request, EntityManagerInterface $entityManager): Response
-    {
-        $typeOf = new TypeOfCompany();
-
-        $form = $this->createForm(TypeOfCompaniesType::class, $typeOf);
+        $form = $this->createForm(CompanyType::class, $company);
         $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()){ 
-            $entityManager->persist($typeOf);
-            $entityManager->flush(); 
+        if ($form->isSubmitted() && $form->isValid()) {
+            $entityManager->flush();
 
-            $this->addFlash('success', 'The new type of company have been added !');
-
-            return $this->redirectToRoute('app_company_typeOf'); 
+            return $this->redirectToRoute('app_company_index', [], Response::HTTP_SEE_OTHER);
         }
 
-        return $this->render('company/addCompanyTypeOf.html.twig', [ // a modifier
-            'controller_name' => 'HomePageController',
-            'formTypeOf'=> $form->createView(),
+        return $this->render('company/edit.html.twig', [
+            'company' => $company,
+            'form' => $form,
         ]);
     }
 
-    #[Route('/company/typeOf/{id}', name: 'app_company_typeOf_show', methods: ['GET'])]
-    public function showTypeOf(TypeOfCompany $typeOf): Response
+    #[Route('/{id}', name: 'app_company_delete', methods: ['POST'])]
+    public function delete(Request $request, Company $company, EntityManagerInterface $entityManager): Response
     {
-        return $this->render('company/showTypeOf.html.twig', [
-            'typeOf' => $typeOf,
-        ]);
+        if ($this->isCsrfTokenValid('delete'.$company->getId(), $request->getPayload()->getString('_token'))) {
+            $entityManager->remove($company);
+            $entityManager->flush();
+        }
+
+        return $this->redirectToRoute('app_company_index', [], Response::HTTP_SEE_OTHER);
     }
-
-    #[Route('/company/typeOfDelete/{id}', name: 'app_company_typeOf_delete')]
-    public function deleteCompaniesTypeOf(TypeOfCompany $typeOf, EntityManagerInterface $entityManager): Response
-    {
-        $entityManager->remove($typeOf);
-        $entityManager->flush(); 
-
-        $this->addFlash('danger', 'The type of company have been deleted.');
-
-        return $this->redirectToRoute('app_company_typeOf'); 
-        
-    }
-
 }
