+++ src/Controller/EmployeeController.php
@@ -3,7 +3,7 @@
 namespace App\Controller;
 
 use App\Entity\Employee;
-use App\Form\EmployeesType;
+use App\Form\EmployeeType;
 use App\Repository\EmployeeRepository;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
@@ -11,43 +11,38 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
 
+#[Route('/employee')]
 final class EmployeeController extends AbstractController
 {
-    #[Route('/employee', name: 'app_employee')]
-    public function index(EmployeeRepository $employeeRepo): Response
+    #[Route(name: 'app_employee_index', methods: ['GET'])]
+    public function index(EmployeeRepository $employeeRepository): Response
     {
-        $employees = $employeeRepo->findAll();
-    
         return $this->render('employee/index.html.twig', [
-            'controller_name' => 'EmployeeController',
-            'employees'=> $employees,
+            'employees' => $employeeRepository->findAll(),
         ]);
     }
 
-    #[Route('/employeeNew', name: 'app_employee_new')]
-    public function addEmployee(Request $request, EntityManagerInterface $entityManager): Response
+    #[Route('/new', name: 'app_employee_new', methods: ['GET', 'POST'])]
+    public function new(Request $request, EntityManagerInterface $entityManager): Response
     {
         $employee = new Employee();
-
-        $form = $this->createForm(EmployeesType::class, $employee);
+        $form = $this->createForm(EmployeeType::class, $employee);
         $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()){ 
+        if ($form->isSubmitted() && $form->isValid()) {
             $entityManager->persist($employee);
-            $entityManager->flush(); 
+            $entityManager->flush();
 
-            $this->addFlash('success', 'The new employee have been added !');
-
-            return $this->redirectToRoute('app_employee'); 
+            return $this->redirectToRoute('app_employee_index', [], Response::HTTP_SEE_OTHER);
         }
 
-        return $this->render('employee/addEmployee.html.twig', [ // a modifier
-            'controller_name' => 'HomePageController',
-            'form'=> $form->createView(),
+        return $this->render('employee/new.html.twig', [
+            'employee' => $employee,
+            'form' => $form,
         ]);
     }
 
-    #[Route('/employee/{id}', name: 'app_employee_show', methods: ['GET'])]
+    #[Route('/{id}', name: 'app_employee_show', methods: ['GET'])]
     public function show(Employee $employee): Response
     {
         return $this->render('employee/show.html.twig', [
@@ -55,15 +50,32 @@ final class EmployeeController extends AbstractController
         ]);
     }
 
-    #[Route('/employeeDelete/{id}', name: 'app_employee_delete')]
-    public function deleteCompanies(Employee $employee, EntityManagerInterface $entityManager): Response
+    #[Route('/{id}/edit', name: 'app_employee_edit', methods: ['GET', 'POST'])]
+    public function edit(Request $request, Employee $employee, EntityManagerInterface $entityManager): Response
     {
-        $entityManager->remove($employee);
-        $entityManager->flush(); 
+        $form = $this->createForm(EmployeeType::class, $employee);
+        $form->handleRequest($request);
 
-        $this->addFlash('danger', 'The employee have been deleted.');
+        if ($form->isSubmitted() && $form->isValid()) {
+            $entityManager->flush();
 
-        return $this->redirectToRoute('app_employee'); 
-        
+            return $this->redirectToRoute('app_employee_index', [], Response::HTTP_SEE_OTHER);
+        }
+
+        return $this->render('employee/edit.html.twig', [
+            'employee' => $employee,
+            'form' => $form,
+        ]);
+    }
+
+    #[Route('/{id}', name: 'app_employee_delete', methods: ['POST'])]
+    public function delete(Request $request, Employee $employee, EntityManagerInterface $entityManager): Response
+    {
+        if ($this->isCsrfTokenValid('delete'.$employee->getId(), $request->getPayload()->getString('_token'))) {
+            $entityManager->remove($employee);
+            $entityManager->flush();
+        }
+
+        return $this->redirectToRoute('app_employee_index', [], Response::HTTP_SEE_OTHER);
     }
 }
